refactor: use new LabelFormat class to manage formatting in labels better
(cherry picked from commit a7651a942da33919f659b0518eef85aed9eb6570)

+++ src/main/java/net/cyvfabric/event/events/ParkourTickListener.java
@@ -56,7 +56,7 @@ public class ParkourTickListener {
     private static boolean grindStarted = false;
 
     public static float last45 = 0;
-    public static double lastTurning = 0;
+    public static float lastTurning = 0;
 
     public static int sidestep = 0; //0 = wad 1 = wdwa
     public static int sidestepTime = -1;
